pamac install отсутсвующие

+++ upgrade.sh
@@ -1,8 +1,41 @@
 #!/bin/bash
-if [[ ! -z "$(find /var/lib/clamav/daily.cvd -type f -mtime +6)" ]]; then echo -e "\n"; 
-  echo -e "База clamav создана более недели назад!"; echo -e "\n"; stat /var/lib/clamav/daily.cvd | grep Модифицирован ; 
-  echo -e "\n"; read -n 1 -p "Обновить базы антивируса clamav? [y/N]: " clupdate;
-  if [[ "$clupdate" = [yYlLдД] ]]; then echo -e "\n"; sudo /home/kostya/my_scripts/update_clamav.sh; fi
+echo -e "Этот скрипт проверяет наличие обновлений и обновляет систему с помощью pamac, yay и paru."
+echo -e "Для полноценной работы скрипта необходимо установить следующие пакеты: "
+echo -e "clamav, timeshift, timeshift-autosnap-manjaro,yay, paru, meld, needrestart и rkhunter."
+echo -e "Скрипт будет работать и без них, только с ограниченной функциональностью."
+echo -e "\n"; read -n 1 -p "Установить отсутствующие пакеты? [y/N]: " inst;
+if [[ "$inst" = [yYlLдД] ]]; then 
+  pamac install --no-confirm clamav timeshift timeshift-autosnap-manjaro yay meld needrestart rkhunter 
+  pamac build --no-confirm paru-bin 
+  if [ ! -f $HOME/my_scripts/rkhunter.sh ]; then 
+    mkdir -p $HOME/my_scripts
+    touch $HOME/my_scripts/rkhunter.sh
+    echo "#!/bin/bash " >> $HOME/my_scripts/rkhunter.sh
+    echo "sudo rkhunter --check --skip-keypress --update --report-warnings-only 2> /dev/null " >> $HOME/my_scripts/rkhunter.sh
+    chmod +x $HOME/my_scripts/rkhunter.sh
+  fi
+  if [ ! -f $HOME/my_scripts/update_clamav.sh ]; then
+    mkdir -p $HOME/my_scripts
+    touch $HOME/my_scripts/update_clamav.sh
+    echo "#!/bin/bash " >> $HOME/my_scripts/update_clamav.sh
+    echo "systemctl stop clamav-freshclam " >> $HOME/my_scripts/update_clamav.sh
+    echo "if [ -f /var/lib/clamav/freshclam.dat ]; then rm -f /var/lib/clamav/freshclam.dat; fi " >> $HOME/my_scripts/update_clamav.sh
+    echo "if [ -f /var/lib/clamav/main.cvd ]; then rm -f /var/lib/clamav/*.cvd; fi " >> $HOME/my_scripts/update_clamav.sh
+    echo "wget https://packages.microsoft.com/clamav/main.cvd -O /var/lib/clamav/main.cvd " >> $HOME/my_scripts/update_clamav.sh
+    echo "wget https://packages.microsoft.com/clamav/daily.cvd -O /var/lib/clamav/daily.cvd " >> $HOME/my_scripts/update_clamav.sh
+    echo "wget https://packages.microsoft.com/clamav/bytecode.cvd -O /var/lib/clamav/bytecode.cvd " >> $HOME/my_scripts/update_clamav.sh
+    echo "stat /var/lib/clamav/daily.cvd | grep Модифицирован " >> $HOME/my_scripts/update_clamav.sh
+    chmod +x $HOME/my_scripts/update_clamav.sh
+  fi
+fi
+package="clamav"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+if [ -n "${check}" ] ; 
+  then
+    if [[ ! -z "$(find /var/lib/clamav/daily.cvd -type f -mtime +6)" ]]; then echo -e "\n"; 
+      echo -e "База clamav создана более недели назад!"; echo -e "\n"; stat /var/lib/clamav/daily.cvd | grep Модифицирован ; 
+      echo -e "\n"; read -n 1 -p "Обновить базы антивируса clamav? [y/N]: " clupdate;
+      if [[ "$clupdate" = [yYlLдД] ]]; then echo -e "\n"; sudo $HOME/my_scripts/update_clamav.sh; fi
+    fi
 fi
 # ---------------------------------------------------------------------------------------------
 # Удаление блокировки баз при ее наличии
@@ -12,10 +45,18 @@ if [[ -f /var/lib/pacman/db.lck ]]; then sudo rm /var/lib/pacman/db.lck; fi
 #if [[ "$cupdate" = [yYlLдД] ]]; then echo -e "\n"; pamac checkupdates -a; fi
 echo -e "\n"; echo -e "Проверка наличия обновлений:"; echo -e "\n"; pamac checkupdates -a
 # ---------------------------------------------------------------------------------------------
-bekaplast=$(find /mnt/sdb/sdb6/timeshift/snapshots -mindepth 1 -maxdepth 1 -printf '%P\n' | sort -r | head -n 1)
-echo -e "\n"; echo -e "Последний бэкап timeshift сделан: " $bekaplast ;
-echo -e "\n"; read -n 1 -p "Сделать бэкап timeshift перед обновлением? [y/N]: " bekap; 
-if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=true/skipAutosnap=false/g' /etc/timeshift-autosnap.conf; fi
+package="timeshift"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+if [ -n "${check}" ] ; 
+  then
+    bekaplast=$(find /mnt/sdb/sdb6/timeshift/snapshots -mindepth 1 -maxdepth 1 -printf '%P\n' | sort -r | head -n 1)
+    echo -e "\n"; echo -e "Последний бэкап timeshift сделан: " $bekaplast ;
+    package="timeshift-autosnap-manjaro"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+    if [ -n "${check}" ] ; 
+      then
+        echo -e "\n"; read -n 1 -p "Сделать бэкап timeshift перед обновлением? [y/N]: " bekap; 
+        if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=true/skipAutosnap=false/g' /etc/timeshift-autosnap.conf; fi
+    fi
+fi
 # echo -e "\n"; read -n 1 -p "Обновить установленные пакеты? [Y/n]: " update; 
 # if [[ "$update" = "" || "$update" = [yYlLдД] ]]; then 
 # pamac upgrade --forse-refresh; 
@@ -34,13 +75,17 @@ if [[ "$updrep" = [yYlLдД] ]]; then
   # ---------------------------------------------------------------------------------------------
   echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
   while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
-  echo -e "\n"; read -n 1 -p "Обновить пакеты из репозиториев через AURхелперы? [y/N]: " upda; 
-  if [[ "$upda" = [yYlLдД] ]]; then
-    echo -e "\n"; read -n 1 -p "Обновить через yay? [y/N]: " yayupd;
-    if [[ "$yayupd" = [yYlLдД] ]]; then echo -e "\n"; yay -Syyuu --repo | tee $HOME/upgrade.yay; fi
-    echo -e "\n"; read -n 1 -p "Обновить через paru? [y/N]: " parupd;
-    if [[ "$parupd" = [yYlLдД] ]]; then echo -e "\n"; paru -Syyuu --repo | tee $HOME/upgrade.paru; fi
-    # if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=false/skipAutosnap=true/g' /etc/timeshift-autosnap.conf; fi
+  package="yay"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+  if [ -n "${check}" ] ; 
+    then
+      echo -e "\n"; read -n 1 -p "Обновить пакеты из репозиториев через AURхелперы? [y/N]: " upda; 
+      if [[ "$upda" = [yYlLдД] ]]; then
+        echo -e "\n"; read -n 1 -p "Обновить через yay? [y/N]: " yayupd;
+        if [[ "$yayupd" = [yYlLдД] ]]; then echo -e "\n"; yay -Syyuu --repo | tee $HOME/upgrade.yay; fi
+        echo -e "\n"; read -n 1 -p "Обновить через paru? [y/N]: " parupd;
+        if [[ "$parupd" = [yYlLдД] ]]; then echo -e "\n"; paru -Syyuu --repo | tee $HOME/upgrade.paru; fi
+        # if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=false/skipAutosnap=true/g' /etc/timeshift-autosnap.conf; fi
+      fi
   fi
   # echo -e "\n";
   # ---------------------------------------------------------------------------------------------
@@ -54,23 +99,37 @@ if [[ "$updrep" = [yYlLдД] ]]; then
   if compgen -G "$HOME/upgrade.*" > /dev/null; then 
     echo -e "\n"; read -n 1 -p "Сравнить конфиги pacnew? [Y/n]: " diff;
     if [[ "$diff" = "" || "$diff" = [yYlLдД] ]]; then 
-      echo -e "\n"; read -n 1 -p "Сравнить в meld(графика)? [Y/n]: " difft;
-      if [[ "$difft" = "" || "$difft" = [yYlLдД] ]]; 
-        then echo -e "\n"; sudo DIFFPROG=meld pacdiff; 
-        else echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff; 
+      package="meld"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+      if [ -n "${check}" ] ; 
+        then
+          echo -e "\n"; read -n 1 -p "Сравнить в meld(графика)? [Y/n]: " difft;
+          if [[ "$difft" = "" || "$difft" = [yYlLдД] ]]; 
+            then echo -e "\n"; sudo DIFFPROG=meld pacdiff; 
+            else echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff; 
+          fi
+        else
+          echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff;
       fi
     fi
     # Конец условия Сравнить конфиги pacnew? 
-    echo -e "\n"; read -n 1 -p "Проверить сервисы для перезапуска? [Y/n]: " restart;
-    if [[ "$restart" = "" || "$restart" = [yYlLдД] ]]; then
-      echo -e "\n"; sudo systemctl daemon-reload; sudo needrestart -u NeedRestart::UI::stdio -r i;  
+    package="needrestart"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+    if [ -n "${check}" ] ; 
+      then
+        echo -e "\n"; read -n 1 -p "Проверить сервисы для перезапуска? [Y/n]: " restart;
+        if [[ "$restart" = "" || "$restart" = [yYlLдД] ]]; then
+          echo -e "\n"; sudo systemctl daemon-reload; sudo needrestart -u NeedRestart::UI::stdio -r i;  
+        fi
     fi
     echo -e "\n"; read -n 1 -p "Проверить, есть ли лишние модули ядра? [y/N]: " kerny; 
     if [[ "$kerny" = [yYlLдД] ]]; then
       echo -e "\n"; echo "В системе установлены следующие ядра:"
       pacman -Q | grep -E "linux[0-9]{2}(\s|[0-9])[^-]"
       echo -e "\n"; echo "Возможно необходимо почистить каталог /usr/lib/modules/"
-      cd /usr/lib/modules/; gksu dbus-run-session thunar /usr/lib/modules/ 2> /dev/null ;
+      package="thunar"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+      if [ -n "${check}" ] ; 
+        then
+        cd /usr/lib/modules/; gksu dbus-run-session thunar /usr/lib/modules/ 2> /dev/null ;
+      fi
     fi
     echo -e "\n"; read -n 1 -p "Проверить, пакеты для пересборки? [y/N]: " pac; 
     if [[ "$pac" = [yYlLдД] ]]; then echo -e "\n";
@@ -90,12 +149,16 @@ if [[ "$updrep" = [yYlLдД] ]]; then
       fi
     fi
     # запуск rkhunter --propupd после изменения конфигурационных файлов или обновления ОС
-    echo -e "\n"; read -n 1 -p "Создать базу данных для rkhunter и выполнить проверку? [y/N]: " rkh;
-    if [[ "$rkh" = [yYlLдД] ]]; then echo -e "\n";
-      sudo rkhunter --propupd 2> /dev/null
-      /home/kostya/my_scripts/rkhunter.sh ; 
-      echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
-      while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
+    package="rkhunter"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+    if [ -n "${check}" ] ; 
+      then
+        echo -e "\n"; read -n 1 -p "Создать базу данных для rkhunter и выполнить проверку? [y/N]: " rkh;
+        if [[ "$rkh" = [yYlLдД] ]]; then echo -e "\n";
+          sudo rkhunter --propupd 2> /dev/null
+          $HOME/my_scripts/rkhunter.sh ; 
+          echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
+          while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
+        fi
     fi
   fi
 fi
@@ -117,12 +180,16 @@ if [[ "$updaur" = [yYlLдД] ]]; then
   fi  
   echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
   while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
-  echo -e "\n"; read -n 1 -p "Обновить пакеты из AUR через AURхелперы? [y/N]: " upda; 
-  if [[ "$upda" = [yYlLдД] ]]; then  
-    echo -e "\n"; read -n 1 -p "Обновить через yay? [y/N]: " yayupd;
-    if [[ "$yayupd" = [yYlLдД] ]]; then echo -e "\n"; yay -Syyu --aur | tee $HOME/upgrade.yay; fi
-    echo -e "\n"; read -n 1 -p "Обновить через paru? [y/N]: " parupd;
-    if [[ "$parupd" = [yYlLдД] ]]; then echo -e "\n"; paru -Syyu --aur | tee $HOME/upgrade.paru; fi
+  package="yay"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+  if [ -n "${check}" ] ; 
+    then
+      echo -e "\n"; read -n 1 -p "Обновить пакеты из AUR через AURхелперы? [y/N]: " upda; 
+      if [[ "$upda" = [yYlLдД] ]]; then  
+        echo -e "\n"; read -n 1 -p "Обновить через yay? [y/N]: " yayupd;
+        if [[ "$yayupd" = [yYlLдД] ]]; then echo -e "\n"; yay -Syyu --aur | tee $HOME/upgrade.yay; fi
+        echo -e "\n"; read -n 1 -p "Обновить через paru? [y/N]: " parupd;
+        if [[ "$parupd" = [yYlLдД] ]]; then echo -e "\n"; paru -Syyu --aur | tee $HOME/upgrade.paru; fi
+      fi
   fi
   if [[ -f $HOME/upgrade.pamac ]]; then if cat $HOME/upgrade.pamac | grep 'Нет заданий.'; then rm $HOME/upgrade.pamac; fi; fi
   if [[ -f $HOME/upgrade.yay ]]; then if cat $HOME/upgrade.yay | grep 'there is nothing to do'; then rm $HOME/upgrade.yay; fi; fi
@@ -131,16 +198,26 @@ if [[ "$updaur" = [yYlLдД] ]]; then
   if compgen -G "$HOME/upgrade.*" > /dev/null; then 
     echo -e "\n"; read -n 1 -p "Сравнить конфиги pacnew? [y/N]: " diff;
     if [[ "$diff" = [yYlLдД] ]]; then 
-      echo -e "\n"; read -n 1 -p "Сравнить в meld(графика)? [Y/n]: " difft;
-      if [[ "$difft" = "" || "$difft" = [yYlLдД] ]]; 
-        then echo -e "\n"; sudo DIFFPROG=meld pacdiff; 
-        else echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff; 
+      package="meld"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+      if [ -n "${check}" ] ; 
+        then
+          echo -e "\n"; read -n 1 -p "Сравнить в meld(графика)? [Y/n]: " difft;
+          if [[ "$difft" = "" || "$difft" = [yYlLдД] ]]; 
+            then echo -e "\n"; sudo DIFFPROG=meld pacdiff; 
+            else echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff; 
+          fi
+        else
+          echo -e "\n"; sudo DIFFPROG=vimdiff pacdiff;
       fi
     fi
     # Конец условия Сравнить конфиги pacnew? 
-    echo -e "\n"; read -n 1 -p "Проверить сервисы для перезапуска? [y/N]: " restart;
-    if [[ "$restart" = [yYlLдД] ]]; then
-      echo -e "\n"; sudo systemctl daemon-reload; sudo needrestart -u NeedRestart::UI::stdio -r i;  
+    package="needrestart"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+    if [ -n "${check}" ] ; 
+      then
+        echo -e "\n"; read -n 1 -p "Проверить сервисы для перезапуска? [y/N]: " restart;
+        if [[ "$restart" = [yYlLдД] ]]; then
+          echo -e "\n"; sudo systemctl daemon-reload; sudo needrestart -u NeedRestart::UI::stdio -r i;  
+        fi
     fi
     echo -e "\n"; read -n 1 -p "Проверить, пакеты для пересборки? [y/N]: " pac; 
     if [[ "$pac" = [yYlLдД] ]]; then echo -e "\n";
@@ -161,18 +238,26 @@ if [[ "$updaur" = [yYlLдД] ]]; then
       fi
     fi
     # запуск rkhunter --propupd после изменения конфигурационных файлов или обновления ОС
-    echo -e "\n"; read -n 1 -p "Создать базу данных для rkhunter и выполнить проверку? [y/N]: " rkh; 
-    if [[ "$rkh" = [yYlLдД] ]]; then echo -e "\n";
-      sudo rkhunter --propupd 2> /dev/null
-      /home/kostya/my_scripts/rkhunter.sh ; 
-      echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
-      while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
+    package="rkhunter"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+    if [ -n "${check}" ] ; 
+      then
+        echo -e "\n"; read -n 1 -p "Создать базу данных для rkhunter и выполнить проверку? [y/N]: " rkh; 
+        if [[ "$rkh" = [yYlLдД] ]]; then echo -e "\n";
+          sudo rkhunter --propupd 2> /dev/null
+          $HOME/my_scripts/rkhunter.sh ; 
+          echo -e "\n"; echo "Нажмите любую клавишу, чтобы продолжить"
+          while true; do read -t 1 variable <&1 ; if [ $? = 0 ] ; then break ; else notify-send -t 600 -i face-plain "   ВНИМАНИЕ! Обновление  " "   Требует <b>Вмешательства</b>  " ; canberra-gtk-play -i dialog-warning ; fi ;  done
+        fi
     fi
   fi
 fi
 # Конец условия Необходимости постобработки после обновления AUR -------------------------------------------------
 # Конец условия Обновить установленные пакеты?
-if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=false/skipAutosnap=true/g' /etc/timeshift-autosnap.conf; fi
+package="timeshift-autosnap-manjaro"; check="$(pacman -Qs --color always "${package}" | grep "local" | grep "${package}")";
+if [ -n "${check}" ] ; 
+  then
+    if [[ "$bekap" = [yYlLдД] ]]; then sudo sed -i 's/skipAutosnap=false/skipAutosnap=true/g' /etc/timeshift-autosnap.conf; fi
+fi
 echo -e "\n";
 # Удаление логов ------------------------------------------------------------------------------
 if [[ -f $HOME/upgrade.paru ]]; then rm $HOME/upgrade.paru; fi
